requirements.txt
asgiref==3.3.4
backcall==0.2.0
certifi==2020.6.20
charset-normalizer==2.0.6
colorama==0.4.3
coverage==5.3
cycler==0.10.0
decorator==4.4.2
Django==3.2.4
django-crispy-forms==1.13.0
django-filter==21.1
django-taggit==1.5.1
djangorestframework==3.12.4
idna==3.2
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
kiwisolver==1.2.0
Markdown==3.3.4
matplotlib==3.3.2
networkx==2.5.1
numpy==1.19.2
parso==0.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.7
psycopg2==2.9.1
psycopg2-binary==2.9.1
pydot==1.4.2
Pygments==2.9.0
pyparsing==3.0.0a2
python-dateutil==2.8.1
pytz==2021.1
requests==2.26.0
scipy==1.7.2
six==1.15.0
sqlparse==0.4.1
stripe==2.60.0
traitlets==5.0.4
urllib3==1.26.7
wcwidth==0.2.5
